cleanup and javadoc
