--- cloud-paper/src/main/java/org/incendo/cloud/paper/sender/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Built in CommandSender mappings for Modern Paper.
- */
-package org.incendo.cloud.paper.sender;
