+++ cloud-paper/src/main/java/org/incendo/cloud/paper/util/sender/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Built in command sender {@link org.incendo.cloud.paper.util.sender.Source types} and
+ * {@link org.incendo.cloud.paper.util.sender.PaperSimpleSenderMapper mappings} for modern Paper.
+ */
+package org.incendo.cloud.paper.util.sender;
